Add files via upload
This code is just basically made for testing and creating person groups and assigning persons to them. I.e, by using Azure api provided by Microsoft. We can train its model for different people and each person can be assigned a specific group so that he belongs to that and only that group.